src: fix parameter printf type

+++ src/datawizard/malloc.c
@@ -125,7 +125,7 @@ int starpu_malloc_flags(void **A, size_t dim, int flags)
 			{
 				size_t freed;
 				size_t reclaim = 2 * dim;
-				_STARPU_DEBUG("There is not enough memory left, we are going to reclaim %ld\n", reclaim);
+				_STARPU_DEBUG("There is not enough memory left, we are going to reclaim %ld\n", (long)reclaim);
 				_STARPU_TRACE_START_MEMRECLAIM(STARPU_MAIN_RAM,0);
 				freed = _starpu_memory_reclaim_generic(STARPU_MAIN_RAM, 0, reclaim);
 				_STARPU_TRACE_END_MEMRECLAIM(STARPU_MAIN_RAM,0);
